feat: add OurOfferSectionTitle component
